requirements.txt
openai
requests
schedule
python-dotenv